remove unused explicit dependency rules for crti/crtn
notes by maintainer:

commit 2f853dd6b9a95d5b13ee8f9df762125e0588df5d added these rules
because the new system for handling arch-provided replacement files
introduced for out-of-tree builds did not apply to the crt tree.

commit 63bcda4d8f4074e9d92ae156afd0dced6e64eb65 later adapted the
makefile logic so that the crt and ldso trees go through the same
replacement logic as everything else, but failed to remove the
explicit rules that assumed the arch would always provide asm
replacements.

in addition to cleaning things up, removing these spurious rules
allows crti/crtn asm to be omitted by an arch (thereby using the empty
C files instead) if they are not needed.